requirements.txt
# Build tools (required for distutils in Python 3.12+)
setuptools==68.2.2

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# OCR and image processing (lightweight)
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow==10.1.0
requests==2.31.0

# Data processing and Excel export
pandas==2.1.3
openpyxl==3.1.2
numpy==1.26.2

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# OpenAI integration (compatible version)
openai==0.28.1

# Optional: EasyOCR (can be heavy, comment out if needed)
# easyocr==1.7.1

# Optional: PyTorch dependencies (very heavy, comment out if using only OCR.space)
# torch==2.1.0
# torchvision==0.16.0

# Optional: PaddleOCR (recommended for container deployment - free and good accuracy)
# paddlepaddle==2.6.1
# paddleocr==2.8.1

# Testing and development (optional for production)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Security (Phase 5B.1: Authentication)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Database (optional)
# alembic==1.12.1
# sqlalchemy==2.0.23
# aiosqlite==0.19.0

# Google Cloud Vision API
google-cloud-vision==3.7.2

# Google Document AI API (for receipt processing)
google-cloud-documentai==2.20.0

# Visualization (optional for production)
# matplotlib==3.8.0
# seaborn==0.12.2
# scikit-learn==1.3.2